requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
future==0.18.2
html5lib==1.1
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.6
PyPDF2==1.26.0
pyrsistent==0.18.1
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.5
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
